+++ src/api/resources/users/resources/rewards/client/requests/index.ts
@@ -1,2 +1,3 @@
 export type { GetLocationsByUserRequest } from "./GetLocationsByUserRequest.js";
+export type { GetOffersByPlacementRequest } from "./GetOffersByPlacementRequest.js";
 export type { GetOffersByUserRequest } from "./GetOffersByUserRequest.js";