+++ tests/wire/users/rewards.test.ts
@@ -206,6 +206,267 @@ describe("RewardsClient", () => {
         }).rejects.toThrow(KardApi.UnauthorizedError);
     });
 
+    test("placementOffers (1)", async () => {
+        const server = mockServerPool.createServer();
+        const client = new KardApiClient({
+            maxRetries: 0,
+            clientId: "client_id",
+            clientSecret: "client_secret",
+            environment: server.baseUrl,
+        });
+
+        const rawResponseBody = {
+            data: [
+                {
+                    type: "standardOffer",
+                    relationships: {
+                        category: {
+                            data: [
+                                { id: "id", type: "category" },
+                                { id: "id", type: "category" },
+                            ],
+                        },
+                    },
+                    id: "id",
+                    attributes: {
+                        terms: "terms",
+                        maxRedemptions: 1,
+                        name: "name",
+                        purchaseChannel: ["INSTORE", "INSTORE"],
+                        userReward: { type: "FLAT", value: 1.1 },
+                        assets: [
+                            { type: "type", url: "url", alt: "alt" },
+                            { type: "type", url: "url", alt: "alt" },
+                        ],
+                        startDate: "2024-01-15T09:30:00Z",
+                        expirationDate: "2024-01-15T09:30:00Z",
+                        isTargeted: true,
+                        minTransactionAmount: { type: "CENTS", value: 1 },
+                        maxTransactionAmount: { type: "CENTS", value: 1 },
+                        minRewardAmount: { type: "CENTS", value: 1 },
+                        maxRewardAmount: { type: "CENTS", value: 1 },
+                        websiteUrl: "websiteUrl",
+                        description: "description",
+                        components: {
+                            shortDescription: "shortDescription",
+                            longDescription: "longDescription",
+                            baseReward: "baseReward",
+                            boostedReward: "boostedReward",
+                            cta: {
+                                buttonText: "buttonText",
+                                buttonStyle: "PRIMARY",
+                                action: { url: "url", method: "method" },
+                                startIcon: "startIcon",
+                            },
+                            tags: ["tags", "tags"],
+                            detailTags: ["detailTags", "detailTags"],
+                            logoFlare: { borderColor: "PRIMARY", badge: { icon: "icon", position: "TOP_RIGHT" } },
+                            progressBar: {
+                                total: 1,
+                                currentProgress: 1,
+                                label: "label",
+                                segmented: true,
+                                labels: { details: {}, default: {} },
+                            },
+                        },
+                    },
+                },
+                {
+                    type: "standardOffer",
+                    relationships: {
+                        category: {
+                            data: [
+                                { id: "id", type: "category" },
+                                { id: "id", type: "category" },
+                            ],
+                        },
+                    },
+                    id: "id",
+                    attributes: {
+                        terms: "terms",
+                        maxRedemptions: 1,
+                        name: "name",
+                        purchaseChannel: ["INSTORE", "INSTORE"],
+                        userReward: { type: "FLAT", value: 1.1 },
+                        assets: [
+                            { type: "type", url: "url", alt: "alt" },
+                            { type: "type", url: "url", alt: "alt" },
+                        ],
+                        startDate: "2024-01-15T09:30:00Z",
+                        expirationDate: "2024-01-15T09:30:00Z",
+                        isTargeted: true,
+                        minTransactionAmount: { type: "CENTS", value: 1 },
+                        maxTransactionAmount: { type: "CENTS", value: 1 },
+                        minRewardAmount: { type: "CENTS", value: 1 },
+                        maxRewardAmount: { type: "CENTS", value: 1 },
+                        websiteUrl: "websiteUrl",
+                        description: "description",
+                        components: {
+                            shortDescription: "shortDescription",
+                            longDescription: "longDescription",
+                            baseReward: "baseReward",
+                            boostedReward: "boostedReward",
+                            cta: {
+                                buttonText: "buttonText",
+                                buttonStyle: "PRIMARY",
+                                action: { url: "url", method: "method" },
+                                startIcon: "startIcon",
+                            },
+                            tags: ["tags", "tags"],
+                            detailTags: ["detailTags", "detailTags"],
+                            logoFlare: { borderColor: "PRIMARY", badge: { icon: "icon", position: "TOP_RIGHT" } },
+                            progressBar: {
+                                total: 1,
+                                currentProgress: 1,
+                                label: "label",
+                                segmented: true,
+                                labels: { details: {}, default: {} },
+                            },
+                        },
+                    },
+                },
+            ],
+            links: { self: "self", prev: "prev", next: "next" },
+            included: [
+                { attributes: { name: "Arts & Entertainment" }, id: "id", type: "category" },
+                { attributes: { name: "Arts & Entertainment" }, id: "id", type: "category" },
+            ],
+            meta: {
+                availableCategories: [
+                    { attributes: { name: "Arts & Entertainment" }, id: "id", type: "category" },
+                    { attributes: { name: "Arts & Entertainment" }, id: "id", type: "category" },
+                ],
+            },
+        };
+
+        server
+            .mockEndpoint()
+            .get("/v2/issuers/organizationId/users/userId/placements/placementId/offers")
+            .respondWith()
+            .statusCode(200)
+            .jsonBody(rawResponseBody)
+            .build();
+
+        const response = await client.users.rewards.placementOffers("organizationId", "userId", "placementId");
+        expect(response).toEqual(rawResponseBody);
+    });
+
+    test("placementOffers (2)", async () => {
+        const server = mockServerPool.createServer();
+        const client = new KardApiClient({
+            maxRetries: 0,
+            clientId: "client_id",
+            clientSecret: "client_secret",
+            environment: server.baseUrl,
+        });
+
+        const rawResponseBody = {
+            errors: [
+                { status: "status", title: "title", detail: "detail" },
+                { status: "status", title: "title", detail: "detail" },
+            ],
+        };
+
+        server
+            .mockEndpoint()
+            .get("/v2/issuers/organizationId/users/userId/placements/placementId/offers")
+            .respondWith()
+            .statusCode(500)
+            .jsonBody(rawResponseBody)
+            .build();
+
+        await expect(async () => {
+            return await client.users.rewards.placementOffers("organizationId", "userId", "placementId");
+        }).rejects.toThrow(KardApi.InternalServerError);
+    });
+
+    test("placementOffers (3)", async () => {
+        const server = mockServerPool.createServer();
+        const client = new KardApiClient({
+            maxRetries: 0,
+            clientId: "client_id",
+            clientSecret: "client_secret",
+            environment: server.baseUrl,
+        });
+
+        const rawResponseBody = {
+            errors: [
+                { status: "status", title: "title", detail: "detail" },
+                { status: "status", title: "title", detail: "detail" },
+            ],
+        };
+
+        server
+            .mockEndpoint()
+            .get("/v2/issuers/organizationId/users/userId/placements/placementId/offers")
+            .respondWith()
+            .statusCode(400)
+            .jsonBody(rawResponseBody)
+            .build();
+
+        await expect(async () => {
+            return await client.users.rewards.placementOffers("organizationId", "userId", "placementId");
+        }).rejects.toThrow(KardApi.InvalidRequest);
+    });
+
+    test("placementOffers (4)", async () => {
+        const server = mockServerPool.createServer();
+        const client = new KardApiClient({
+            maxRetries: 0,
+            clientId: "client_id",
+            clientSecret: "client_secret",
+            environment: server.baseUrl,
+        });
+
+        const rawResponseBody = {
+            errors: [
+                { status: "status", title: "title", detail: "detail" },
+                { status: "status", title: "title", detail: "detail" },
+            ],
+        };
+
+        server
+            .mockEndpoint()
+            .get("/v2/issuers/organizationId/users/userId/placements/placementId/offers")
+            .respondWith()
+            .statusCode(404)
+            .jsonBody(rawResponseBody)
+            .build();
+
+        await expect(async () => {
+            return await client.users.rewards.placementOffers("organizationId", "userId", "placementId");
+        }).rejects.toThrow(KardApi.DoesNotExistError);
+    });
+
+    test("placementOffers (5)", async () => {
+        const server = mockServerPool.createServer();
+        const client = new KardApiClient({
+            maxRetries: 0,
+            clientId: "client_id",
+            clientSecret: "client_secret",
+            environment: server.baseUrl,
+        });
+
+        const rawResponseBody = {
+            errors: [
+                { status: "status", title: "title", detail: "detail" },
+                { status: "status", title: "title", detail: "detail" },
+            ],
+        };
+
+        server
+            .mockEndpoint()
+            .get("/v2/issuers/organizationId/users/userId/placements/placementId/offers")
+            .respondWith()
+            .statusCode(401)
+            .jsonBody(rawResponseBody)
+            .build();
+
+        await expect(async () => {
+            return await client.users.rewards.placementOffers("organizationId", "userId", "placementId");
+        }).rejects.toThrow(KardApi.UnauthorizedError);
+    });
+
     test("locations (1)", async () => {
         const server = mockServerPool.createServer();
         const client = new KardApiClient({